backend/requirements.txt
fastapi
sqlalchemy
openai
nltk
uvicorn
pydantic
python-dotenv
fuzzywuzzy